+++ src/restaurant/service/restaurant.service.ts
@@ -4,22 +4,22 @@ import {
   CreateRestaurantInput,
   UpdateRestaurantInput,
 } from '../dto/restaurant-input';
-import { UserRepository } from 'src/user/repository/user.repository';
 import { v4 as uuidv4 } from 'uuid';
+import { OwnerRepository } from 'src/user/repository/owner.repository';
 
 @Injectable()
 export class RestaurantService {
   constructor(
     private readonly restaurantRepository: RestaurantRepository,
-    private readonly userRepository: UserRepository,
+    private readonly ownerRepository: OwnerRepository,
   ) {}
 
   async createRestaurant(
-    ownerId: string,
+    userId: string,
     { name, address }: CreateRestaurantInput,
   ) {
-    const owner = await this.userRepository.getUserById(ownerId);
-    if (!owner) {
+    const ownerId = await this.ownerRepository.getOwnerIdByUserId(userId);
+    if (!ownerId) {
       throw new Error('Owner not found');
     }
     await this.restaurantRepository.saveRestaurant(
@@ -39,13 +39,21 @@ export class RestaurantService {
   }
 
   async updateRestaurant(
+    userId: string,
     restaurantId: string,
     { name, address }: UpdateRestaurantInput,
   ) {
+    const ownerId = await this.ownerRepository.getOwnerIdByUserId(userId);
+    if (!ownerId) {
+      throw new Error('Owner not found');
+    }
     const restaurant =
       await this.restaurantRepository.getRestaurantById(restaurantId);
     if (!restaurant) {
       throw new Error('Restaurant not found');
+    }
+    if (restaurant.ownerId !== ownerId) {
+      throw new Error('You are not the owner of this restaurant');
     }
 
     await this.restaurantRepository.saveRestaurant(
@@ -56,7 +64,19 @@ export class RestaurantService {
     );
   }
 
-  async deleteResetaurant(restaurantId: string) {
+  async deleteResetaurant(userId: string, restaurantId: string) {
+    const ownerId = await this.ownerRepository.getOwnerIdByUserId(userId);
+    if (!ownerId) {
+      throw new Error('Owner not found');
+    }
+    const restaurant =
+      await this.restaurantRepository.getRestaurantById(restaurantId);
+    if (!restaurant) {
+      throw new Error('Restaurant not found');
+    }
+    if (restaurant.ownerId !== ownerId) {
+      throw new Error('You are not the owner of this restaurant');
+    }
     await this.restaurantRepository.deleteRestaurant(restaurantId);
   }
 }
